# api/openapi/myaccount/v1/openapi.yaml
openapi: "3.1.0"
info:
  title: MyAccount BFF API
  version: "1.0.0"
  description: |
    Backend-for-Frontend API for the MyAccount SPA.
    All /api/* endpoints require an authenticated session cookie.
    Auth endpoints handle OIDC login/logout flows.

servers:
  - url: /

tags:
  - name: auth
    description: Authentication flow endpoints
  - name: profile
    description: User profile management
  - name: linked-accounts
    description: Federated identity link management
  - name: sessions
    description: Active session management
  - name: account
    description: Account-level operations

paths:
  /auth/login:
    get:
      operationId: authLogin
      tags: [auth]
      summary: Initiate OIDC login
      description: Redirects user-agent to accounts-idp authorization endpoint.
      parameters:
        - name: return_to
          in: query
          schema:
            type: string
          description: URL to redirect to after successful login
      responses:
        "302":
          description: Redirect to IdP authorization endpoint

  /auth/callback:
    get:
      operationId: authCallback
      tags: [auth]
      summary: OIDC callback
      description: |
        Handles the authorization code callback from accounts-idp.
        Exchanges the code for tokens, creates a session, sets the cookie,
        and redirects to the SPA.
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect to SPA after successful authentication
        "400":
          description: Invalid callback parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/logout:
    post:
      operationId: authLogout
      tags: [auth]
      summary: Logout and destroy session
      responses:
        "200":
          description: Successfully logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/session:
    get:
      operationId: getSession
      tags: [auth]
      summary: Get current session info
      description: Returns basic info about the current session for the SPA header.
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/v1/profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          $ref: "#/components/responses/Unauthorized"

    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/v1/linked-accounts:
    get:
      operationId: listLinkedAccounts
      tags: [linked-accounts]
      summary: List linked identity providers
      responses:
        "200":
          description: List of linked accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkedAccountList"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/v1/linked-accounts/{id}:
    delete:
      operationId: unlinkAccount
      tags: [linked-accounts]
      summary: Unlink an identity provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Successfully unlinked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Cannot unlink the only remaining identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/v1/sessions:
    get:
      operationId: listSessions
      tags: [sessions]
      summary: List active sessions
      responses:
        "200":
          description: List of active sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionList"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/v1/sessions/{id}:
    delete:
      operationId: revokeSession
      tags: [sessions]
      summary: Revoke a specific session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Session revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/v1/account:
    delete:
      operationId: deleteAccount
      tags: [account]
      summary: Permanently delete account
      description: Permanently and irreversibly deletes the user's account.
      responses:
        "204":
          description: Account deleted
        "401":
          $ref: "#/components/responses/Unauthorized"

components:
  schemas:
    Profile:
      type: object
      required: [user_id, email, email_verified, given_name, family_name, locale, created_at, updated_at]
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        email_verified:
          type: boolean
        given_name:
          type: string
        family_name:
          type: string
        picture:
          type: string
        locale:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    UpdateProfileRequest:
      type: object
      properties:
        given_name:
          type: string
          minLength: 1
          maxLength: 100
        family_name:
          type: string
          minLength: 1
          maxLength: 100
        picture:
          type: string
        locale:
          type: string
          pattern: "^[a-z]{2}(-[A-Z]{2})?$"

    LinkedAccount:
      type: object
      required: [id, provider, external_sub, linked_at]
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
        external_sub:
          type: string
        linked_at:
          type: string
          format: date-time

    LinkedAccountList:
      type: object
      required: [linked_accounts]
      properties:
        linked_accounts:
          type: array
          items:
            $ref: "#/components/schemas/LinkedAccount"

    Session:
      type: object
      required: [session_id, client_id, scopes, auth_time, expires_at, created_at]
      properties:
        session_id:
          type: string
          format: uuid
        client_id:
          type: string
        scopes:
          type: array
          items:
            type: string
        auth_time:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time

    SessionList:
      type: object
      required: [sessions]
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/Session"

    SessionInfo:
      type: object
      required: [authenticated, user_id, email, given_name, family_name]
      properties:
        authenticated:
          type: boolean
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        given_name:
          type: string
        family_name:
          type: string
        picture:
          type: string

    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string

  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"