# server/bff/gen/myaccount/v1/types.gen.go
// Package myaccountv1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package myaccountv1

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Error defines model for Error.
type Error struct {
	Code    string `json:"code"`
	Message string `json:"message"`
}

// LinkedAccount defines model for LinkedAccount.
type LinkedAccount struct {
	ExternalSub string             `json:"external_sub"`
	Id          openapi_types.UUID `json:"id"`
	LinkedAt    time.Time          `json:"linked_at"`
	Provider    string             `json:"provider"`
}

// LinkedAccountList defines model for LinkedAccountList.
type LinkedAccountList struct {
	LinkedAccounts []LinkedAccount `json:"linked_accounts"`
}

// MessageResponse defines model for MessageResponse.
type MessageResponse struct {
	Message string `json:"message"`
}

// Profile defines model for Profile.
type Profile struct {
	CreatedAt     time.Time           `json:"created_at"`
	Email         openapi_types.Email `json:"email"`
	EmailVerified bool                `json:"email_verified"`
	FamilyName    string              `json:"family_name"`
	GivenName     string              `json:"given_name"`
	Locale        string              `json:"locale"`
	Picture       *string             `json:"picture,omitempty"`
	UpdatedAt     time.Time           `json:"updated_at"`
	UserId        openapi_types.UUID  `json:"user_id"`
}

// Session defines model for Session.
type Session struct {
	AuthTime  time.Time          `json:"auth_time"`
	ClientId  string             `json:"client_id"`
	CreatedAt time.Time          `json:"created_at"`
	ExpiresAt time.Time          `json:"expires_at"`
	Scopes    []string           `json:"scopes"`
	SessionId openapi_types.UUID `json:"session_id"`
}

// SessionInfo defines model for SessionInfo.
type SessionInfo struct {
	Authenticated bool                `json:"authenticated"`
	Email         openapi_types.Email `json:"email"`
	FamilyName    string              `json:"family_name"`
	GivenName     string              `json:"given_name"`
	Picture       *string             `json:"picture,omitempty"`
	UserId        openapi_types.UUID  `json:"user_id"`
}

// SessionList defines model for SessionList.
type SessionList struct {
	Sessions []Session `json:"sessions"`
}

// UpdateProfileRequest defines model for UpdateProfileRequest.
type UpdateProfileRequest struct {
	FamilyName *string `json:"family_name,omitempty"`
	GivenName  *string `json:"given_name,omitempty"`
	Locale     *string `json:"locale,omitempty"`
	Picture    *string `json:"picture,omitempty"`
}

// BadRequest defines model for BadRequest.
type BadRequest = Error

// NotFound defines model for NotFound.
type NotFound = Error

// Unauthorized defines model for Unauthorized.
type Unauthorized = Error

// AuthCallbackParams defines parameters for AuthCallback.
type AuthCallbackParams struct {
	Code  string `form:"code" json:"code"`
	State string `form:"state" json:"state"`
}

// AuthLoginParams defines parameters for AuthLogin.
type AuthLoginParams struct {
	// ReturnTo URL to redirect to after successful login
	ReturnTo *string `form:"return_to,omitempty" json:"return_to,omitempty"`
}

// UpdateProfileJSONRequestBody defines body for UpdateProfile for application/json ContentType.
type UpdateProfileJSONRequestBody = UpdateProfileRequest