# web/apps/myaccount/src/features/security/hooks/useSessions.ts
import { useQuery } from '@tanstack/react-query';
import { api } from '@/shared/api/client';
import type { components } from '@/shared/api/schema';

export type Session = components['schemas']['Session'];

export const useSessions = () => {
  return useQuery({
    queryKey: ['sessions'],
    queryFn: async () => {
      const { data, error } = await api.GET('/api/v1/sessions');
      if (error) throw error;
      return data;
    },
  });
};